requirements.txt
pyTelegramBotAPI==4.4.1
PyQt5~=5.15.6
PyYAML~=6.0